+++ wooqi/src/sequencer_features.py
@@ -10,94 +10,21 @@ Sequencer features
 import os
 import time
 import pytest
+from itertools import product
 from wooqi.src import global_var
 
 
-def get_uuts(pytest_metafunc, cpt=None):
+def get_uuts(test_name):
     """
     Return uut and uut2 according to test
     """
-    if cpt is None:
-        test_name = pytest_metafunc.function.__name__
-    else:
-        test_name = '{}_{}'.format(pytest_metafunc.function.__name__, cpt)
     uut = []
     uut2 = []
-    if 'uut' in pytest_metafunc.fixturenames:
-        uut += global_var['config'].uut(test_name)
-    if 'uut2' in pytest_metafunc.fixturenames:
-        uut2 += global_var['config'].uut2(test_name)
+    if 'uut' in global_var['config'].file_config[test_name]:
+        uut = global_var['config'].uut(test_name)
+    if 'uut2' in global_var['config'].file_config[test_name]:
+        uut2 = global_var['config'].uut2(test_name)
     return uut, uut2
-
-
-def init_loop_option(pytest_metafunc, cpt=None):
-    """
-    Generate test call according to loop option
-    """
-    if global_var['config'].loop_infos() is not None:
-        loop_tests = global_var['config'].loop_infos()[0]
-        loop_iter = int(global_var['config'].loop_infos()[1])
-        if cpt is None:
-            test_name = pytest_metafunc.function.__name__
-        else:
-            test_name = '{}_{}'.format(pytest_metafunc.function.__name__, cpt)
-            # addcall function has already been used in the first loop
-            loop_iter -= 1
-        test_order = global_var['config'].file_config[test_name]['test_order']
-        first_loop_test_order = global_var['config'].file_config[loop_tests[0]]["test_order"]
-        last_loop_test_order = global_var['config'].file_config[loop_tests[1]]["test_order"]
-        if first_loop_test_order <= test_order <= last_loop_test_order:
-            pytest_metafunc.function.__dict__["add_call"] = True
-            pytest_metafunc.function.__dict__["loop"] = True
-            # Add call for each test performed after the first
-            for nb in range(0, loop_iter):
-                pytest_metafunc.addcall()
-
-
-def configure_reruns_timeout_features(items):
-    """
-    Configure the tests with the reruns and timeout features
-    according to the config test file
-    """
-    for item in items:
-        test_name = item.name.split("[")[0]
-        if global_var['config'].exist("{}_0".format(test_name)):
-            test_name = "{}_{}".format(test_name, item.name.split("[")[1][0])
-        if global_var['config'].exist(test_name):
-            if global_var['config'].reruns(test_name) is not None:
-                item.keywords.__dict__["reruns"] = 0
-                item.add_marker(pytest.mark.flaky(
-                    reruns=global_var['config'].reruns(test_name)))
-            if global_var['config'].timeout(test_name) is not None:
-                item.add_marker(pytest.mark.timeout(
-                    global_var['config'].timeout(test_name)))
-
-
-def item_name_analyze(item_name):
-    """
-    Return a basic test name and parameters
-    """
-    item_args = []
-    if '[' not in item_name:
-        test_basic_name = item_name
-    else:
-        # Create list with all args
-        test_basic_name, parameters = item_name.split("[", 1)
-        parameters = parameters.rsplit("]", 1)[0]
-        # First character can to be '-' but not to be a separator
-        previous_char_separator = True
-        parameter = ''
-        for index, char in enumerate(parameters):
-            if previous_char_separator or char != '-':
-                parameter += char
-                previous_char_separator = False
-            else:
-                item_args.append(parameter)
-                parameter = ''
-                previous_char_separator = True
-        item_args.append(parameter)
-
-    return test_basic_name, item_args
 
 
 def item_loop_option_analyse(item, test_name, first_arg):
@@ -151,61 +78,121 @@ def filter_order_tests(config, items):
     """
     Filter and re-order the tests according to the config test file
     """
-    list_temp = []
-    list_loop_temp = []
-    for item in items:
-        basic_test_name, item_args = item_name_analyze(item.name)
-        if not global_var['config'].exist(basic_test_name) and\
-           not global_var['config'].exist('{}_0'.format(basic_test_name)):
-            # Test not used
-            continue
-
-        if item_args:
-            first_arg = item_args[0]
-        else:
-            first_arg = None
-        config_test_name, iter_number = item_loop_option_analyse(item, basic_test_name, first_arg)
-
-        # If len(item_args) is 1, item_args[0] is a uut and he exists
-        if "add_call" in item.keywords.__dict__["_markers"].keys():
-            uut_index = 1
-        else:
-            uut_index = 0
-
-        if len(item_args) > uut_index:
-            # Verify uut
-            if item_args[uut_index] not in global_var['config'].uut(config_test_name):
-                # uut not used for this test
+    pytest_tests = {}
+    file_config = global_var['config'].file_config
+    for session in file_config:
+        if session.startswith('test_') or session.startswith('action_'):
+            if session == 'test_info':
                 continue
-            # Verify uut2
-            if len(item_args) > 2:
-                if item_args[uut_index + 1] not in global_var['config'].uut2(config_test_name):
-                    # uut2 not used for this test
-                    continue
 
-        test_order = int(global_var['config'].file_config[config_test_name]['test_order'])
-        if iter_number != 0:
-            # Create second list only with test in loop option after the first iteration
-            if not list_loop_temp:
-                test_order_last_loop_test = global_var['config'].file_config[
-                    global_var['config'].loop_infos()[0][1]]["test_order"]
-                # Add a fake item who the test_oder is just after the last test of looping option
-                list_temp.append([test_order_last_loop_test + 0.5, 'LOOP_TEST'])
-            list_loop_temp.append([iter_number, test_order, item])
-        else:
-            list_temp.append([test_order, item])
+            if 'test_info' in file_config and 'loop_iter' in file_config['test_info']:
+                if 'wooqi_loop_iter' not in file_config[session]:
+                    loop_tests, iter_number = global_var['config'].loop_infos()
+                    loop_test_start = global_var['config'].loop_infos()[0][0]
+                    iter_number = file_config[loop_test_start]['test_order']
+                    # All tests before the loop start use the first iteration number
+                    if file_config[session]['test_order'] < iter_number:
+                        iter_number = 0
+                    # All tests after loop option use the last iteration number
+                    else:
+                        iter_number = global_var['config'].loop_infos()[1] - 1
+                # Else: For tests in the loop, the iter_number is defined after
+            else:
+                # Without loop option, all tests are the same iter_number level
+                iter_number = 0
 
-    # Order items
-    list_temp.sort(key=lambda x: x[0])
+            if len(session.split('-')) == 0:
+                test_name = session
+            else:
+                test_name = session.replace('-', '[')
+
+            if 'wooqi_loop_iter' in file_config[session]:
+                loop_iter = file_config[session]['wooqi_loop_iter']
+                loop_iter = range(loop_iter)
+            else:
+                loop_iter = []
+
+            uut, uut2 = get_uuts(session)
+
+            if uut2:
+                if loop_iter:
+                    list_param = list(product(loop_iter, uut, uut2))
+                else:
+                    list_param = list(product(uut, uut2))
+            elif uut:
+                if loop_iter:
+                    list_param = list(product(loop_iter, uut))
+                else:
+                    list_param = uut
+            elif loop_iter:
+                list_param = loop_iter
+            else:
+                list_param = []
+
+            test_order = file_config[session]['test_order']
+            if list_param:
+                if '[' not in test_name:
+                    test_name = '{}['.format(test_name)
+                for params in list_param:
+                    if type(params) is tuple:
+                        concatenate_params = ''
+                        for param in params:
+                            concatenate_params += '{}-'.format(param)
+                        concatenate_params = concatenate_params[:-1]
+                    else:
+                        concatenate_params = params
+                    test_name_with_param = '{}-{}]'.format(
+                        test_name, concatenate_params).replace('[-', '[')
+
+                    if loop_iter:
+                        if type(params) is tuple:
+                            iter_number = params[0]
+                        else:
+                            iter_number = params
+                        pytest_tests[test_name_with_param] = {'session': session,
+                                                              'test_order': test_order,
+                                                              'iter_number': iter_number}
+                    else:
+                        pytest_tests[test_name_with_param] = {'session': session,
+                                                              'test_order': test_order,
+                                                              'iter_number': iter_number}
+            else:
+                if '[' in test_name:
+                    test_name += ']'
+                pytest_tests[test_name] = {'session': session,
+                                           'test_order': test_order,
+                                           'iter_number': iter_number}
+
+    # Link test_name with the function
+    for item in items:
+        if item.name in pytest_tests:
+            pytest_tests[item.name]['item'] = item
+
+            session_name = pytest_tests[item.name]['session']
+            # To manage reruns option
+            if 'reruns' in file_config[session_name]:
+                pytest_tests[item.name]['item'].keywords.__dict__["reruns"] = 0
+                pytest_tests[item.name]['item'].add_marker(
+                        pytest.mark.flaky(reruns=global_var['config'].reruns(session_name)))
+            # To manage timeout option
+            if 'timeout' in file_config[session_name]:
+                pytest_tests[item.name]['item'].add_marker(
+                        pytest.mark.timeout(global_var['config'].timeout(session_name)))
+
+    # Check if all action/test defined in the configuration file exist
+    for test_name in pytest_tests:
+        if 'item' not in pytest_tests[test_name]:
+            msg = '{} not found, ' \
+                'please verify the test definition and your configuration file'.format(test_name)
+            print(msg)
+            raise Exception(msg)
+
+    # Create an ordered list
+    ordered_list = sorted(pytest_tests, key=lambda k: (
+            pytest_tests[k]['iter_number'], pytest_tests[k]['test_order'], k))
     items_temp = []
-    for item in list_temp:
-        if str(item[1]).count('LOOP_TEST'):
-            # Insert all tests of loop option instead of fake item
-            list_loop_temp.sort(key=lambda x: (x[0], x[1]))
-            for loop_item in list_loop_temp:
-                items_temp.append(loop_item[2])
-        else:
-            items_temp.append(item[1])
+    for test_name in ordered_list:
+        items_temp.append(pytest_tests[test_name]['item'])
     items[:] = items_temp
 
     # Manage --failed-first option to pytest-cache for wooqi
@@ -230,56 +217,42 @@ def rerun_sequence_since_the_fail(config, items):
     cache_path = os.path.abspath('.cache/v/cache/{}'.format(serial_number))
     if os.path.isfile(cache_path):
         # Found the name of test failed
-        cache_file = open(cache_path, 'r')
-        file_read = cache_file.read()
+        with open(cache_path, 'r') as f:
+            file_read = f.read()
 
         sequence_name = file_read.split('\n', 1)[0]
         if config.getoption('--seq-config') != sequence_name:
             return
 
         item_fail_name = file_read.split('::')[1].split('"')[0]
-        if '[' in item_fail_name:
+        if '[' not in item_fail_name:
+            test_failed = item_fail_name
+        else:
             test_failed, option = item_fail_name.split('[', 1)
             option = option.split('-')[0].split(']')[0]
-            if global_var['config'].exist('{}_0'.format(test_failed)):
-                # First option is not uut
-                item_fail_name = '{}[{}'.format(test_failed, option)
-                test_failed = '{}_{}'.format(test_failed, option)
+
+            if '{}-{}'.format(test_failed, option) in global_var['config'].file_config:
+                item_fail_name = '{}-{}'.format(test_failed, option)
+                test_failed = '{}[{}-'.format(test_failed, option)
             else:
-                item_fail_name = test_failed
+                item_fail_name = test_failed.split('[')[0]
 
             # Verify loop option
-            if global_var['config'].loop_infos() is not None:
-                test_order_loop_start = global_var['config'].file_config[
-                    global_var['config'].loop_infos()[0][0]]["test_order"]
-                test_order_loop_stop = global_var['config'].file_config[
-                    global_var['config'].loop_infos()[0][1]]["test_order"]
-                # If test fail in loop, restart the loop
-                if not global_var['config'].exist(test_failed) or \
-                    (test_order_loop_start <=
-                     global_var['config'].file_config[test_failed]['test_order'] <=
-                     test_order_loop_stop):
-                    test_failed = global_var['config'].loop_infos()[0][0]
-                    test_failed, option = test_failed.rsplit('_', 1)
-                    if global_var['config'].exist('{}_{}'.format(test_failed, option)) and \
-                            global_var['config'].exist('{}_0'.format(test_failed)):
-                        # First option is not uut
-                        option = option.split(']', 1)[0].split('-')[0]
-                        item_fail_name = '{}[{}'.format(test_failed, option)
-                        test_failed = '{}_{}'.format(test_failed, option)
-                    else:
-                        item_fail_name = '{}_{}'.format(test_failed, option)
-        else:
-            test_failed = item_fail_name
+            if 'wooqi_loop_iter' in global_var['config'].file_config[item_fail_name]:
+                item_fail_name = global_var['config'].loop_infos()[0][0]
+
+                test_failed = '{}'.format(item_fail_name.split('-')[0])
+                if '-' in item_fail_name:
+                    test_failed = '{}[{}-'.format(test_failed, item_fail_name.split('-')[1])
 
         # Verify if one test is required
-        try:
-            test_required = global_var['config'].file_config[test_failed]['test_required']
-            if test_required is not None:
-                # Found item_fail_name
-                item_fail_name = test_required
-        except KeyError:
-            pass
+        if 'test_required' in global_var['config'].file_config[item_fail_name]:
+            # Found item_fail_name
+            item_fail_name = global_var['config'].file_config[item_fail_name]['test_required']
+
+            test_failed = '{}'.format(item_fail_name.split('-')[0])
+            if '-' in item_fail_name:
+                test_failed = '{}[{}-'.format(test_failed, item_fail_name.split('-')[1])
 
         # Authorize all tests after found the first test failed
         first_test_failed_found = False
@@ -287,69 +260,49 @@ def rerun_sequence_since_the_fail(config, items):
         for item in items:
             if first_test_failed_found:
                 items_temp.append(item)
-            elif item_fail_name in str(item):
-                if '[' not in str(item_fail_name) or \
-                        str(item).split(item_fail_name)[1][0] in [']', '-']:
-                    first_test_failed_found = True
-                    items_temp.append(item)
+            elif test_failed in item.name or (
+                    test_failed.endswith('-') and '{}]'.format(test_failed[:-1]) == item.name):
+                first_test_failed_found = True
+                items_temp.append(item)
         items[:] = items_temp
 
 
-def postfail_feature_management(item, skip, loop):
+def postfail_feature_management(item, skip):
     """
     Manage the postfail feature according to the config test file
     """
     if not skip:
         return
 
-    # Name defined in config file
-    item_name, item_args = item_name_analyze(item.name)
-    if 'loop' in item.keywords.__dict__["_markers"].keys():
-        item_name = item_loop_option_analyse(item, item_name, item_args[0])[0]
-    elif global_var['config'].exist(item_name + "_0"):
-        item_name = '%s_%s' % (item_name, item_args[0])
+    skip_reason = ""
+    file_config = global_var['config'].file_config
 
-    post_fail = global_var['config'].post_fail(item_name)
-    if post_fail is not None:
-        if global_var['config'].post_fail(item_name) == "next_step":
-            pass
-
-        elif "test" in post_fail or "action" in post_fail:
-            skip_reason = ""
-            for session_item in item.session.items:
-                if "[" not in session_item.name:
-                    session_item_name = session_item.name
-                else:
-                    session_item_name, session_item_args = item_name_analyze(session_item.name)
-                    if global_var['config'].exist(session_item_name + "_0"):
-                        session_item_name = '%s_%s' % (session_item_name, session_item_args[0])
-
-                if post_fail in session_item_name:
-                    break
-
-                if 'loop' in item.keywords.__dict__["_markers"].keys():
-                    if item_name.rsplit('_', 1)[0] != session_item_name.rsplit('_', 1)[0]:
-                        session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))
-                # if same base name, verify if it is the same iteration to authorize another uut
-                    elif item_args[0] != session_item_args[0]:
-                        session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))
-                elif item_name != session_item_name:
-                    session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))
+    # test_name according in file_config
+    test_name = item.name.split('[')[0]
+    if test_name not in file_config:
+        call_number = item.name.split('[')[1].split('-')[0].replace(']', '')
+        base_name = '{}[{}-'.format(test_name, call_number)
+        test_name = '{}-{}'.format(test_name, call_number)
     else:
-        skip_reason = ""
-        for session_item in item.session.items:
-            if "[" not in session_item.name:
-                session_item_name = session_item.name
-            else:
-                session_item_name, session_item_args = item_name_analyze(session_item.name)
-                if global_var['config'].exist(session_item_name + "_0"):
-                    session_item_name = '%s_%s' % (session_item_name, session_item_args[0])
+        base_name = '{}['.format(test_name)  # base_name is the item name without parameters
 
-            if 'loop' in item.keywords.__dict__["_markers"].keys():
-                if item_name.rsplit('_', 1)[0] != session_item_name.rsplit('_', 1)[0]:
-                    session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))
-                # if same base name, verify if it is the same iteration to authorize another uut
-                elif item_args[0] != session_item_args[0]:
-                    session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))
-            elif item_name != session_item_name:
-                session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))
+    if 'wooqi_loop_iter' in file_config[test_name]:
+        base_name += item.name.replace(base_name, '').split('-')[0].replace(']', '')
+
+    post_fail = global_var['config'].post_fail(test_name)
+    if post_fail is not None:
+        if post_fail.lower() == "next_step":
+            return
+        else:
+            if '-' in post_fail:
+                # Adapt post_fail to item.name format
+                post_fail = '{}[{}]'.format(post_fail.split('-')[0], post_fail.split('-')[1])
+
+    for session_item in item.session.items:
+        if post_fail and (post_fail == session_item.name or
+                          post_fail.replace(']', '-') in session_item.name):
+            break
+        elif base_name in session_item.name:
+            continue
+        else:
+            session_item.add_marker(pytest.mark.skipif(skip, reason=skip_reason))